removed authentication middleware and added a middleware to return jwt tokens
